(CDAP-14663) Refactoring LogAppender guice module
- Remove the LoggingModules class
- Use separate guice modules that fits the need
