+++ cdap-app-fabric/src/main/java/co/cask/cdap/app/guice/DistributedProgramContainerModule.java
@@ -50,7 +50,8 @@ import co.cask.cdap.internal.app.runtime.workflow.MessagingWorkflowStateWriter;
 import co.cask.cdap.internal.app.runtime.workflow.WorkflowStateWriter;
 import co.cask.cdap.logging.appender.LogAppender;
 import co.cask.cdap.logging.appender.tms.TMSLogAppender;
-import co.cask.cdap.logging.guice.LoggingModules;
+import co.cask.cdap.logging.guice.KafkaLogAppenderModule;
+import co.cask.cdap.logging.guice.TMSLogAppenderModule;
 import co.cask.cdap.messaging.guice.MessagingClientModule;
 import co.cask.cdap.metadata.MetadataReaderWriterModules;
 import co.cask.cdap.metrics.guice.MetricsClientRuntimeModule;
@@ -197,7 +198,7 @@ public class DistributedProgramContainerModule extends AbstractModule {
 
     modules.add(new LocationRuntimeModule().getDistributedModules());
     modules.add(new KafkaClientModule());
-    modules.add(new LoggingModules().getDistributedModules());
+    modules.add(new KafkaLogAppenderModule());
     modules.add(new DataFabricModules(generateClientId(programRunId, instanceId)).getDistributedModules());
     modules.add(new DataSetsModules().getDistributedModules());
     modules.add(new NamespaceQueryAdminModule());
@@ -205,6 +206,7 @@ public class DistributedProgramContainerModule extends AbstractModule {
   }
 
   private void addIsolatedModules(List<Module> modules) {
+    modules.add(new TMSLogAppenderModule());
     modules.add(new DataSetsModules().getStandaloneModules());
     modules.add(new DataSetServiceModules().getStandaloneModules());
     modules.add(Modules.override(new DataFabricModules().getInMemoryModules()).with(new AbstractModule() {
@@ -226,8 +228,6 @@ public class DistributedProgramContainerModule extends AbstractModule {
     modules.add(new AbstractModule() {
       @Override
       protected void configure() {
-        bind(LogAppender.class).to(TMSLogAppender.class);
-
         // Bind to unsupported/no-op class implementations for features that are not supported in isolated cluster mode
         bind(StreamAdmin.class).to(UnsupportedStreamAdmin.class);
         bind(StreamCoordinatorClient.class).to(InMemoryStreamCoordinatorClient.class);
