+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/runtime/distributed/remote/RemoteExecutionTwillRunnerService.java
@@ -42,6 +42,7 @@ import co.cask.cdap.data2.transaction.Transactions;
 import co.cask.cdap.internal.app.runtime.ProgramOptionConstants;
 import co.cask.cdap.internal.app.runtime.SystemArguments;
 import co.cask.cdap.internal.app.runtime.distributed.ProgramTwillApplication;
+import co.cask.cdap.internal.app.runtime.monitor.RemoteExecutionLogProcessor;
 import co.cask.cdap.internal.app.runtime.monitor.RuntimeMonitor;
 import co.cask.cdap.internal.app.runtime.monitor.RuntimeMonitorClient;
 import co.cask.cdap.internal.app.runtime.monitor.RuntimeMonitorServerInfo;
@@ -49,7 +50,6 @@ import co.cask.cdap.internal.app.runtime.monitor.proxy.MonitorSocksProxy;
 import co.cask.cdap.internal.profile.ProfileMetricService;
 import co.cask.cdap.internal.provision.LocationBasedSSHKeyPair;
 import co.cask.cdap.internal.provision.ProvisioningService;
-import co.cask.cdap.logging.remote.RemoteExecutionLogProcessor;
 import co.cask.cdap.messaging.MessagingService;
 import co.cask.cdap.messaging.context.MultiThreadMessagingContext;
 import co.cask.cdap.proto.id.NamespaceId;
