+++ cdap-master/src/main/java/co/cask/cdap/data/runtime/main/MasterServiceMain.java
@@ -62,7 +62,7 @@ import co.cask.cdap.explore.service.ExploreServiceUtils;
 import co.cask.cdap.hive.ExploreUtils;
 import co.cask.cdap.internal.app.services.AppFabricServer;
 import co.cask.cdap.logging.appender.LogAppenderInitializer;
-import co.cask.cdap.logging.guice.LoggingModules;
+import co.cask.cdap.logging.guice.KafkaLogAppenderModule;
 import co.cask.cdap.master.startup.ServiceResourceKeys;
 import co.cask.cdap.messaging.guice.MessagingClientModule;
 import co.cask.cdap.metrics.guice.MetricsClientRuntimeModule;
@@ -530,7 +530,7 @@ public class MasterServiceMain extends DaemonMain {
     return Guice.createInjector(
       new ConfigModule(cConf, hConf),
       new ZKClientModule(),
-      new LoggingModules().getDistributedModules()
+      new KafkaLogAppenderModule()
     );
   }
 
@@ -553,7 +553,7 @@ public class MasterServiceMain extends DaemonMain {
           bind(LeaderElectionInfoService.class).toInstance(electionInfoService);
         }
       },
-      new LoggingModules().getDistributedModules(),
+      new KafkaLogAppenderModule(),
       new LocationRuntimeModule().getDistributedModules(),
       new IOModule(),
       new ZKDiscoveryModule(),
