+++ cdap-master/src/main/java/co/cask/cdap/data/runtime/main/MetricsProcessorTwillRunnable.java
@@ -32,7 +32,7 @@ import co.cask.cdap.data.runtime.DataFabricModules;
 import co.cask.cdap.data.runtime.DataSetsModules;
 import co.cask.cdap.data2.audit.AuditModule;
 import co.cask.cdap.logging.appender.LogAppenderInitializer;
-import co.cask.cdap.logging.guice.LoggingModules;
+import co.cask.cdap.logging.guice.KafkaLogAppenderModule;
 import co.cask.cdap.messaging.guice.MessagingClientModule;
 import co.cask.cdap.metrics.guice.MetricsClientRuntimeModule;
 import co.cask.cdap.metrics.guice.MetricsProcessorStatusServiceModule;
@@ -110,7 +110,7 @@ public final class MetricsProcessorTwillRunnable extends AbstractMasterTwillRunn
       new MessagingClientModule(),
       new MetricsClientRuntimeModule().getDistributedModules(),
       new MetricsStoreModule(),
-      new LoggingModules().getDistributedModules(),
+      new KafkaLogAppenderModule(),
       new LocationRuntimeModule().getDistributedModules(),
       new NamespaceQueryAdminModule(),
       new DataFabricModules(txClientId).getDistributedModules(),
