+++ cdap-master/src/main/java/co/cask/cdap/data/tools/UpgradeTool.java
@@ -67,7 +67,7 @@ import co.cask.cdap.internal.app.runtime.schedule.queue.JobQueueDataset;
 import co.cask.cdap.internal.app.runtime.schedule.store.ProgramScheduleStoreDataset;
 import co.cask.cdap.internal.app.runtime.schedule.store.ScheduleStoreTableUtil;
 import co.cask.cdap.internal.app.store.DefaultStore;
-import co.cask.cdap.logging.guice.LoggingModules;
+import co.cask.cdap.logging.guice.KafkaLogAppenderModule;
 import co.cask.cdap.logging.meta.LoggingStoreTableUtil;
 import co.cask.cdap.messaging.guice.MessagingClientModule;
 import co.cask.cdap.messaging.store.hbase.HBaseTableFactory;
@@ -232,7 +232,7 @@ public class UpgradeTool {
       new SecureStoreModules().getDistributedModules(),
       new DataFabricModules(UpgradeTool.class.getName()).getDistributedModules(),
       new AppFabricServiceRuntimeModule().getDistributedModules(),
-      new LoggingModules().getDistributedModules(),
+      new KafkaLogAppenderModule(),
       // the DataFabricDistributedModule needs MetricsCollectionService binding
       new MetricsStoreModule(),
       new AbstractModule() {
