--- cdap-watchdog/src/main/java/co/cask/cdap/logging/guice/LoggingModules.java
@@ -1,66 +1,0 @@
-/*
- * Copyright © 2014 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package co.cask.cdap.logging.guice;
-
-import co.cask.cdap.common.runtime.RuntimeModule;
-import co.cask.cdap.logging.appender.LogAppender;
-import co.cask.cdap.logging.appender.kafka.KafkaLogAppender;
-import co.cask.cdap.logging.framework.local.LocalLogAppender;
-import co.cask.cdap.logging.remote.LogAppenderLogProcessor;
-import co.cask.cdap.logging.remote.RemoteExecutionLogProcessor;
-import com.google.inject.AbstractModule;
-import com.google.inject.Module;
-import com.google.inject.Scopes;
-
-/**
- * Injectable modules for logging.
- */
-public class LoggingModules extends RuntimeModule {
-
-  @Override
-  public Module getInMemoryModules() {
-    return new AbstractModule() {
-      @Override
-      protected void configure() {
-        bind(LogAppender.class).to(LocalLogAppender.class).in(Scopes.SINGLETON);
-        bind(RemoteExecutionLogProcessor.class).to(LogAppenderLogProcessor.class).in(Scopes.SINGLETON);
-      }
-    };
-  }
-
-  @Override
-  public Module getStandaloneModules() {
-    return new AbstractModule() {
-      @Override
-      protected void configure() {
-        bind(LogAppender.class).to(LocalLogAppender.class).in(Scopes.SINGLETON);
-        bind(RemoteExecutionLogProcessor.class).to(LogAppenderLogProcessor.class).in(Scopes.SINGLETON);
-      }
-    };
-  }
-
-  @Override
-  public Module getDistributedModules() {
-    return new AbstractModule() {
-      @Override
-      protected void configure() {
-        bind(LogAppender.class).to(KafkaLogAppender.class).in(Scopes.SINGLETON);
-        bind(RemoteExecutionLogProcessor.class).to(LogAppenderLogProcessor.class);
-      }
-    };
-  }
-}
