+++ cdap-watchdog/src/test/java/co/cask/cdap/logging/clean/FileMetadataCleanerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2017 Cask Data, Inc.
+ * Copyright © 2017-2018 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -45,7 +45,7 @@ import co.cask.cdap.logging.LoggingConfiguration;
 import co.cask.cdap.logging.appender.system.CDAPLogAppender;
 import co.cask.cdap.logging.appender.system.LogPathIdentifier;
 import co.cask.cdap.logging.context.LoggingContextHelper;
-import co.cask.cdap.logging.guice.LoggingModules;
+import co.cask.cdap.logging.guice.LocalLogAppenderModule;
 import co.cask.cdap.logging.meta.FileMetaDataReader;
 import co.cask.cdap.logging.meta.FileMetaDataWriter;
 import co.cask.cdap.logging.meta.LoggingStoreTableUtil;
@@ -106,7 +106,7 @@ public class FileMetadataCleanerTest {
       new ConfigModule(cConf, hConf),
       new NonCustomLocationUnitTestModule().getModule(),
       new TransactionModules().getInMemoryModules(),
-      new LoggingModules().getInMemoryModules(),
+      new LocalLogAppenderModule(),
       new DataSetsModules().getInMemoryModules(),
       new SystemDatasetRuntimeModule().getInMemoryModules(),
       new AuthorizationTestModule(),
